feat(maps): Add locked door to Beach map
This commit adds a locked door object to the FOS_Beach1.tmx map.

The locked door is implemented using Tiled's object layers and properties. The "Action" property is set to "Action Message aztecviper.fishofstardew_LockedDoor", which triggers a message when the player interacts with the door. This message is defined in the i18n files and injected via the `StringsFromMaps` data.